backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
django-cors-headers>=4.3
pip install mysqlclient
